third_party/blink/web_tests/external/wpt/client-hints/sec-ch-width-auto-sizes-002.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<meta http-equiv="Delegate-CH" content="Sec-CH-Width">
<title>Tests Sec-CH-Width with auto sizes and multiple srcset values</title>
<link rel="help" href="https://wicg.github.io/responsive-image-client-hints/#sec-ch-width">
<link rel="help" href="https://html.spec.whatwg.org/#sizes-attributes">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
  setup({explicit_done: true});

  function run() {
    test(() => {
      const testImage = document.getElementsByTagName('img')[0];
      assert_equals(testImage.naturalWidth, 2 * 50 * window.devicePixelRatio);
      assert_equals(testImage.naturalHeight, 3 * 50 * window.devicePixelRatio);
    }, 'Sec-CH-Width is set for lazy auto sizes');
    done();
  }
</script>

<div style="width: 100px; height: 100px;">
  <img
    onload="run()"
    loading="lazy"
    sizes="auto"
    srcset="fail-a 49w, resources/2x3-svg-scaled-by-sec-ch-width.py 50w, fail-b 51w"
    width="300"
    height="300"
    style="width: 50%; height: 50%;">
</div>
